// api-gateway/src/common/circuit-breaker/circuit-breaker.interface.ts
export enum CircuitBreakerStateEnum {
    CLOSED = 'CLOSED',
    OPEN = 'OPEN',
    HALF_OPEN = 'HALF_OPEN'
}

export interface CircuitBreakerOptions {
    failureThreshold: number;
    timeout: number;
    resetTimeout: number;
}

export interface CircuitBreakerState {
    state: CircuitBreakerStateEnum;
    failureCount: number;
    lastFailureTime: number;
    nextAttemptTime: number;
}

// <T> generic type being used in data
export interface CircuitBreakerResult<T> {
    success: boolean;
    data?: T;
    error?: Error;
    fromCache?: boolean;
}